feat: mentor disclaimer and styling for form
