+++ apps/site/src/app/apply/hacker/HackerApplication.tsx
@@ -2,11 +2,9 @@
 
 import { useState } from "react";
 
-import RetroWindow from "@/components/RetroWindow/RetroWindow";
-import BaseForm from "@/components/BaseForm/BaseForm";
-import ApplicationForm from "./ApplicationForm";
 import ApplicationsClosed from "./ApplicationsClosed";
 import HackerDisclaimer from "./HackerDisclaimer";
+import HackerForm from "./HackerForm";
 
 import hasDeadlinePassed from "@/lib/utils/hasDeadlinePassed";
 
@@ -14,7 +12,6 @@ import styles from "./HackerApplication.module.scss";
 
 export default function HackerApplication() {
 	const [hasAcceptedDisclaimer, setHasAcceptedDisclaimer] = useState(false);
-	const [page, setPage] = useState<1 | 2>(1);
 	const deadlinePassed = hasDeadlinePassed();
 
 	if (deadlinePassed) return <ApplicationsClosed />;
@@ -27,23 +24,5 @@ export default function HackerApplication() {
 		);
 	}
 
-	const title =
-		page === 1 ? "Hacker Application" : "Hacker Application (Continued)";
-
-	return (
-		<div className={styles.page}>
-			<div className={styles.windowWrapper}>
-				<RetroWindow title={title}>
-					<BaseForm
-						applyPath="/api/user/apply"
-						applicationType="Hacker"
-						className={styles.form}
-						hideSubmit={page === 1}
-					>
-						<ApplicationForm page={page} onPageChange={setPage} />
-					</BaseForm>
-				</RetroWindow>
-			</div>
-		</div>
-	);
+	return <HackerForm />;
 }
