--- apps/site/src/app/apply/hacker/PixelArt.module.scss
@@ -1,274 +1,0 @@
-@use "zothacks-theme" as theme;
-
-.wrapper {
-	max-width: 28rem;
-	margin: 0 auto;
-	padding: 1.5rem 1rem 4rem;
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	gap: 1rem;
-	color: theme.$white;
-}
-
-.heading {
-	font-size: 1.75rem;
-	font-weight: 700;
-	margin: 0 0 0.25rem;
-	text-align: center;
-}
-
-// Grid
-.grid {
-	display: grid;
-	grid-template-columns: repeat(8, 1fr);
-	gap: 0;
-	width: 16rem;
-	height: 16rem;
-	user-select: none;
-	border: 2px solid rgba(255, 255, 255, 0.12);
-	box-shadow: 0 4px 18px -4px rgba(0, 0, 0, 0.4) inset;
-	background: rgba(255, 255, 255, 0.04);
-}
-
-.pixel {
-	width: 2rem;
-	height: 2rem;
-	box-sizing: border-box;
-	border: 1px solid rgba(255, 255, 255, 0.08);
-	transition:
-		background-color 0.08s linear,
-		box-shadow 0.15s ease;
-	cursor: pointer;
-	&:hover,
-	&:active {
-		box-shadow: 0 0 0 2px rgba(255, 255, 255, 0.25) inset;
-	}
-}
-
-// Palette
-.palette {
-	display: grid;
-	grid-template-columns: repeat(9, 1fr);
-	grid-template-rows: repeat(3, 1fr);
-	gap: 0.25rem;
-	width: 16rem;
-	height: 6.5rem;
-	justify-content: center;
-}
-
-.swatch {
-	width: 2rem;
-	height: 2rem;
-	border: 2px solid transparent;
-	border-radius: 0.375rem;
-	cursor: pointer;
-	padding: 0;
-	box-shadow: 0 2px 4px rgba(0, 0, 0, 0.4);
-	transition:
-		transform 0.15s ease,
-		box-shadow 0.2s ease,
-		border-color 0.15s;
-}
-
-.dragging .swatch {
-	transform: none !important;
-	pointer-events: none;
-}
-
-.selected {
-	border-color: #fff;
-	box-shadow: 0 0 0 2px rgba(255, 255, 255, 0.35);
-}
-
-// Buttons / Controls
-.controlsRow {
-	display: flex;
-	gap: 0.5rem;
-	width: 100%;
-}
-
-.controlBtn {
-	flex: 1;
-	font: inherit;
-	padding: 0.55rem 0.75rem;
-	border-radius: 0.6rem;
-	border: 1px solid rgba(255, 255, 255, 0.18);
-	background: rgba(255, 255, 255, 0.07);
-	color: theme.$white;
-	cursor: pointer;
-	display: inline-flex;
-	align-items: center;
-	justify-content: center;
-	gap: 0.25rem;
-	transition:
-		background 0.2s ease,
-		border-color 0.2s,
-		transform 0.15s ease;
-	text-align: center;
-	&:hover {
-		background: rgba(255, 255, 255, 0.12);
-	}
-	&:active {
-		transform: scale(0.96);
-	}
-}
-
-.secondaryBtn {
-	@extend .controlBtn;
-}
-
-.dangerBtn {
-	@extend .controlBtn;
-	background: linear-gradient(135deg, #dc2626, #b91c1c);
-	&:hover {
-		background: linear-gradient(135deg, #b91c1c, #991b1b);
-	}
-}
-
-// Modals
-.modal {
-	border: none;
-	padding: 0;
-	background: transparent;
-	&::backdrop {
-		background: rgba(0, 0, 0, 0.6);
-		backdrop-filter: blur(2px);
-	}
-}
-
-.modalBox {
-	width: min(90vw, 26rem);
-	background: linear-gradient(
-		165deg,
-		rgba(27, 36, 67, 0.9),
-		rgba(60, 80, 147, 0.9)
-	);
-	border: 1px solid rgba(255, 255, 255, 0.15);
-	border-radius: 1rem;
-	padding: 1.25rem 1.25rem 1rem;
-	box-shadow:
-		0 8px 30px -8px rgba(0, 0, 0, 0.55),
-		0 2px 4px rgba(0, 0, 0, 0.4) inset;
-	color: theme.$white;
-	display: flex;
-	flex-direction: column;
-	gap: 0.75rem;
-}
-
-.modalTitle {
-	margin: 0;
-	font-size: 1.15rem;
-	font-weight: 600;
-}
-
-.modalBody {
-	margin: 0;
-	font-size: 0.85rem;
-	line-height: 1.3;
-	opacity: 0.92;
-}
-
-.modalActions {
-	display: flex;
-	gap: 0.5rem;
-	justify-content: flex-end;
-}
-
-// Color classes - 3 rows of 9 colors (light, normal, dark)
-// Row 1: Light variants
-.cWhite {
-	background: #ffffff;
-}
-.cRedLight {
-	background: #fca5a5;
-}
-.cOrangeLight {
-	background: #fdba74;
-}
-.cYellowLight {
-	background: #fde047;
-}
-.cGreenLight {
-	background: #86efac;
-}
-.cBlueLight {
-	background: #93c5fd;
-}
-.cIndigoLight {
-	background: #a5b4fc;
-}
-.cPurpleLight {
-	background: #c4b5fd;
-}
-.cPinkLight {
-	background: #f9a8d4;
-}
-
-// Row 2: Normal variants
-.cGray {
-	background: #9ca3af;
-}
-.cRed {
-	background: #f87171;
-}
-.cOrange {
-	background: #fb923c;
-}
-.cYellow {
-	background: #facc15;
-}
-.cGreen {
-	background: #4ade80;
-}
-.cBlue {
-	background: #60a5fa;
-}
-.cIndigo {
-	background: #818cf8;
-}
-.cPurple {
-	background: #a78bfa;
-}
-.cPink {
-	background: #f472b6;
-}
-
-// Row 3: Dark variants
-.cBlack {
-	background: #374151;
-}
-.cRedDark {
-	background: #dc2626;
-}
-.cOrangeDark {
-	background: #ea580c;
-}
-.cYellowDark {
-	background: #ca8a04;
-}
-.cGreenDark {
-	background: #16a34a;
-}
-.cBlueDark {
-	background: #2563eb;
-}
-.cIndigoDark {
-	background: #4f46e5;
-}
-.cPurpleDark {
-	background: #7c3aed;
-}
-.cPinkDark {
-	background: #db2777;
-}
-
-// Accessibility focus outline
-.swatch:focus-visible,
-.pixel:focus-visible,
-.controlBtn:focus-visible,
-.secondaryBtn:focus-visible,
-.dangerBtn:focus-visible {
-	outline: 2px solid #fff;
-	outline-offset: 2px;
-}
