+++ apps/site/src/app/apply/mentor/MentorApplication.module.scss
@@ -1,0 +1,40 @@
+.page {
+	background: var(--blue-gradient);
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	justify-content: flex-start;
+	padding: 5rem 1rem 3rem;
+}
+
+.windowWrapper {
+	width: 100%;
+	max-width: 1100px;
+}
+
+.disclaimerShell {
+	min-height: calc(100vh - 50px);
+	box-sizing: border-box;
+	justify-content: flex-start;
+	overflow-y: auto;
+	overflow-x: hidden;
+	padding: 5rem 1rem 3rem;
+}
+
+.form {
+	display: flex;
+	flex-direction: column;
+	gap: 1.5rem;
+	padding: 1.5rem;
+	background: #ffffff;
+}
+
+@media (max-width: 600px) {
+	.page {
+		padding: 4rem 0.5rem 2rem;
+	}
+
+	.disclaimerShell {
+		padding: 5rem 1rem 3rem;
+	}
+}
