--- apps/site/src/app/apply/mentor/Mentors.module.scss
@@ -1,13 +1,0 @@
-.background {
-	background: linear-gradient(
-		135deg,
-		rgba(27, 36, 67) 0%,
-		rgba(60, 80, 147) 50%,
-		rgba(42, 80, 75) 100%
-	);
-
-	min-height: 100dvh;
-	padding-bottom: clamp(3rem, 8vh, 6rem);
-	width: 100%;
-	box-sizing: border-box;
-}
